hugo/requirements.txt
astroid==2.1.0
ilio==0.4.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.2.2
python-frontmatter==0.4.4
PyYAML==3.13
six==1.12.0
toml==0.10.0
typed-ast==1.1.1
wrapt==1.10.11
